fix: fix test log
